Extract documentation for qml methods and signals
The documentation of qml methods and signals are found after the mark
$$$methodOrSignalName[overload1]$$$ mark if that method is documented
individually. When a list of methods share the same documentation
this convention does not apply. In this case, we extract the common
after matching <span class="name">methodName paragraph.

Also introduce an helper Element data structure to abstract the item to
be searched in the html document.

Fixes: QTBUG-122122
Change-Id: I7f3c44509234274ebb39b4348fb8f6b3e26ce0b8
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>