Attempt to stabilize tst_how-to-qml's timePicker test
Top-level animations within a transition are run in parallel, which
meant that the NumberAnimations that animated the opacity of the labels
could technically be conflicting (even though it visually looked
correct). This could explain why the test was failing when checking
for the opacity to equal 1, and it was instead 0.96.

Reduce the duration of the animations now that they execute in
succession, as originally intended.

Amends d414c81f38d8fea47d91c9414bdecda9ea7761a1.

Task-number: QTBUG-122679
Pick-to: 6.6
Change-Id: I51c62d8c95d54194fd2fb2273c9644378d4b61e5
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit ee988e556af28042070ceb535f56f9ac58612947)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>