Fix assert in curve renderer on arm64
A left-over development assert would hit on some architectures.

The intention was just to check that some floating point computation
results should be fuzzy-0, but it was sloppily done, mixing floats and
double precisions and the documented non-working qFuzzyCompare'ing
with 0. This made it sensitive to slight differences in precision on
different cpu architectures.

As the assert no longer serves any purpose (the code is now done and
nothing catastrophic will happen anyway if these results somehow be
non-fuzzy-0), just remove it.

Fixes: QTBUG-122676
Pick-to: 6.7
Change-Id: I240135fe60518e1fb5dad8e788aba688c2315f21
Reviewed-by: Eskil Abrahamsen Blomfeldt <eskil.abrahamsen-blomfeldt@qt.io>
Reviewed-by: Paul Olav Tvete <paul.tvete@qt.io>