Hover event delivery: don't modify a list we are iterating
Delivering HoverLeave events might modify or even clear the list of
hoverItems, which invalidates the iterators, and crash.

Iterate over a copy of the list instead, and write the modified
copy back to the hoverItems list if that hasn't been cleared while
delivering events. To prevent that we deliver multiple HoverLeave
events to items, check in the delivery logic that the item's entry
in the map still has a valid ID for the event point. Streamline
that code a bit by preventing multiple lookups in the flatmap, and
instead use an iterator that we can use to update the existing
entry if it's valid.

Pick-to: 6.6 6.5
Fixes: QTBUG-122686
Change-Id: I5483e75884f1ddec37c4e98ecfee35e7596756c5
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit d00e890e5dbcc559bd63ae996e2b31485e3b03bf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>