QQuickRt: allow specifying a viewFormat where applicable
Some of the fromXxxx static constructors now get a third (elsewhere,
second) overload. This has all the Qt 6.8 new functionality merged
into one to prevent the explosion of these functions.
(arraySize for multiview, flags for multisample-resolve (and other
future uses), and a viewFormat in addition to format)

Also order everything in the logical way: first the Qt 6.0 era versions
(that do not have a format argument), then the Qt 6.4 overloads with
format, then the Qt 6.8 overload with the full set of features.

With OpenGL we do not and will not do texture views. Hence no viewFormat
argument. However, there is a version requirement: OpenGL ES 3.0 (or 2.0)
is not sufficient for some of the new features. Given the main customer for
these is Quick 3D XR with devices like the Quest 3 with OpenGL ES 3.2,
this should not be a problem in practice.

D3D11 has no multiview support, hence no arraySize argument. Also no
viewFormat argument since casting is limited with D3D11 (would need
typeless formats, which we do not do). The sRGB problem is still
solvable by passing in a modified format (stripping _SRGB). The same
goes for OpenGL, just adjust the format.

D3D12 and Vulkan have full support for everything, although viewFormat
requires Windows 1703 or newer to be functional. This should be fine
nowadays.

The Metal version also has all the arguments, even though viewFormat
may be ignored for the time being, until support is implemented in
the QRhi backend.

Task-number: QTBUG-122288
Task-number: QTBUG-122614
Change-Id: Ia760ce785b72378aba26695db2a4d57e10eaac90
Reviewed-by: Andy Nichols <andy.nichols@qt.io>