Fall back to retrying with "software" when swapchain fails
The idea being that a swapchain creation failure would end up behaving
the same as as the QRhi::create() failure, in particular on Windows:
for the latter we automatically retry with PreferSwRenderer set (i.e.
requesting WARP), but until now this has not been done upon swapchain
creation failures. (it is not clear under what circumstances we fail
to create a swapchain, but reports from Qt-based applications indicate
that this is happening out in the wild)

If the first swapchain creation attempt fails, do what a device loss
would do: drop everything, incl. the QRhi, and post a special event
to the main thread do request a new round of rendering, upon which
we'll attempt to reinitialize a QRhi and everything else in the
scenegraph. We then recognize the flag and pass it on to QSGRhiSupport
to indicate we want a software-backed adapter / physical device,
if there is one. We won't retry afterwards if that fails too.

This may also correct issues with device loss handling which did not
correctly wake up the main thread.

[ChangeLog][Qt Quick] The fallback to a software rasterizer, if
applicable to the platform and 3D API, is now performed also upon
the first swapchain creation failure. Previously this was only done
if the QRhi initialization failed. Relevant in particular on Windows,
potentially allowing functioning on systems that are incapable of
proper accelerated D3D rendering, but, for whatever reason, do not
fail early on upon the device/context creation, only later at
swapchain creation.

Pick-to: 6.6 6.5
Fixes: QTBUG-109708
Fixes: QTBUG-101200
Change-Id: I76f4aa132361d8f97ec6edfb3bf7806b4ce015b1
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Andy Nichols <andy.nichols@qt.io>
(cherry picked from commit 96299182fbc0b11a89ef153bce5be8268b9d87dd)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>