QQuickMenu: allow non-native menubar in combination with native drop-down menus
The current useNativeMenu() implementation would force all menus
inside a non-native menubar to be non-native as well. But this is
not how it's supposed to work after all. On e.g Windows, the "native"
menubar is old and unmaintained, doesn't look native, and isn't used
by Widgets eiter. Instead, the solution, also for Widgets, is to use
a non-native menubar on Windows (and Linux), but use native menus
(or native windows with non-native contents).

This patch will therefore change the implementation of useNativeMenu()
to allow this combination.

Change-Id: Ie526c73f173f39f813cb0923ad70e517b0648e5e
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>