StackView: Fix and clarify pushItems documentation
The code snippet needs to use pushItems not push (which was a leftover
from before we renamed the function).
Moreover, clarify that items itself has to be an array, and mention the
possibility of passing arguments a bit earlier.

Change-Id: I8c7cf3576c692b2b5f543a706ec61406c37906dc
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
(cherry picked from commit 6406dd28bc39d160f1c87869babc6ed31b04c4df)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>