Binding: fix restore logic
Consider a Binding affecting a property which initially has a value (not
a binding).
When restoring the initial value of a property, we did not remove any
binding that might have been installed on it by Binding. Then, when the
Binding would subsequently get enabled again, it would pick up the still
installed binding instead of the actually restored property.
From that point on, the original value would have been lost.
Fix this by clearing any existing bindings when we restore a value.

Pick-to: 6.7 6.6 6.5
Fixes: QTBUG-122707
Change-Id: Ife7c43c7e799c71257fc04bbd76ff902422bd49d
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Olivier De Cannière <olivier.decanniere@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>