QQuickTableView: remove syncView from destructor
If a TableView has a syncView, we need to ensure that it detaches
itself from the syncView when it's destroyed. Otherwise the
syncView will have a dangling pointer to the deleted TableView
afterwards, which can cause a crash.

Fixes: QTBUG-120760
Pick-to: 6.6 6.5
Change-Id: I4c6acfaa0c623ea43ba8b938585fcd9c9247f66c
Reviewed-by: Santhosh Kumar <santhosh.kumar.selvaraj@qt.io>
(cherry picked from commit e5c031c1ac542be0d2ac7457e145da011fce0013)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>