Warn about unset required properties on composite singletons
So far, we only checked for errors when creating the Component for the
singleton. However, the creation of the singleton should also fail when
some of its (or its subobjects) required properties are not set.

C++ singletons will be handled in a separate change.

[ChangeLog][Potentially Source-Incompatible Changes][QML] Instantiating
singletons with unset required properties will now fail and print a
warning instead of creating the singleton with the properties unset.

Task-number: QTBUG-122784
Pick-to: 6.8 6.7 6.5
Change-Id: I80c8b036cb839b4cc19683e181e2c08f29c2fab5
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>