doc: Clarify that automatic transient parent relies on visual parent
As part of 16023fc77c423a267fcc48894ff942e94cf35b86, the behavior
of declaratively and imperatively setting the visible property of
a window was made consistent, which is an important property of
QML.

Both methods now respect the presence of an Item or Window parent
for the Window, and resolve a possible transient parent from the
parent Item or Window. The transient parent magic behavior is
active as long as the transientParent property is not set by
the user explicitly.

Seeing as transient child windows should not be shown before their
transient parent windows, this means that as long as we can possibly
have a transient parent, we don't show the window.

And since an Item may not initially be part of a scene, it may not
have an initial window, and we have to wait to show the child window.

What this means is that an Item parent hierarchy of a Window needs
to be in a scene, so that the direct Item parent of a Window gets
a window, which we can then use to determine whether to show the
transient child window or not.

This matches the behavior of an Item added as a property to another
item. It will not be rendered until it's included in the scene by
setting the visual parent of the item to some item that's in the
scene.

[ChangeLog][QtQuick][Window] Windows that automatically pick up
a transient parent window from its Item parent will not be shown
unless the item is part of a scene.

Fixes: QTBUG-121197
Change-Id: Iae050cd5f0a4e11ef9c397494dac9bbac8d00cdd
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 6db942a87db172f6d707afa0bfaf1278cb5db965)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>