Clean up and refactor hover event delivery
Replace bool-trap with a dedicated enum to indicate whether hover
should get set or cleared.
Use structured bindings to decompose the std::pair we get from the
flat-map, and assert that the id has the right value (0 when clearing,
non-0 when setting) after delivering the hover event.

Remove the call to clear() on the flat-map of hoverItems in clearHover;
at this point, all items in the map must have an ID of 0, and will be
removed from the map in the next run of deliverHoverEvent. This
way, the hoverItems list is no longer modified while we iterate over
the list of hoverItems, and we can remove the copy of the flatmap
introduced as a low-impact fix in
d00e890e5dbcc559bd63ae996e2b31485e3b03bf.

Change-Id: Ic5fa7cb63ab4ee62f6465e6c126e9109ef28f7e2
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>