tst_qmlls_utils: move test files in own folder
Move .qml test files in own folder. This prepares the code for the commit that
find usages in multiple files, without breaking all existing tests that
do not expect usages to be found in different files.

For example, multiple .qml files are checking for usages of
QQuickItem's "widthChanged()", but do not expect other files to have
usages of widthChanged(). This means that adding a new test file that
contains a widthChanged() might potentially break an unrelated test once
find usages starts finding usages from files in the implicit import
directory, for example.

Task-number: QTBUG-113827
Change-Id: I4170e0abc8b6eb1e5e34b48c99e01a2e77ccc6e8
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>