Fail gracefully when binding an attached property where the object is null
Overlay.overlay is an attached property that holds the window overlay
item, our documentation promises that it can be attached to anything,
but will be null if the thing is not in a window.

When using QQuickView on a QML code that declares an item which
tries to bind something to the Overlay.overlay property, then the overlay
will be null, and creating it will fail because there is also no window
yet. This crashes further down processing when trying to create the
binding.

While this cannot be made to work (modifying the window is not possible
in QML that doesn't declare a Window and is instead loaded into a
QQuickView), it should also not crash.

Add a nullptr check where it did crash, add a test that reproduces the
crash, and with the fix merely spits out a warning.

Fixes: QTBUG-122894
Pick-to: 6.5
Change-Id: Ie4438657a6855b44409edb28431e04eb9875a392
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
Reviewed-by: Kwanghyo Park <kwanghyo.park@qt.io>
(cherry picked from commit bad5a7ac1fe3ec39c6c9a0111f1514177a69315e)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 56de91bec09bb0f14f13b2522efe547ffd86000e)
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>