QML tooling: populate import path with dependency location
Now that we can pass targets to DEPENDENCIES/IMPORTS, we can extract
their output location, and pass it on in the CMake API.
This is currently only possible with non-imported targets, and thus is
only helpful for the case where multiple modules are build as part of a
project, but not for the deployment use case.
The testcase is based on Mitch's example from QTBUG-122986.

Task-number: QTBUG-122986
Change-Id: Ibe264bc3b877a1fe65ef0655fbea5b8ea188cda6
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>