Prevent infinite loop when setting activeFocusItem of the scene's root
When we forceActiveFocus on a QQuickItem we set focus to all ancestor
focus scopes, in their parent scope. That involves setting 'item'
as the subFocusItem of its immediate focus ancestor.
In Qt Quick 3D, where we have different subscenes, we follow the
same approach as mentioned above, but at the same time, we would set the
item that is getting focus as the activeFocusItem of the scene’s root
item and by so doing, we would set the subFocusItem of all focus
ancestors of that item to 'item'. This is not entirely correct since
each focus ancestor's subFocusItem must be its nearest descendant with
focus. But we would then proceed to go up the focus hierarchy of the
item and this time try to set its ancestor as the activeFocusItem of the
scene’s root, "undoing" what we had done in the previous step. This
would cause a lot of unnecessary clearing and setting of the
subFocusItem as well as focusIn/out events, and sometimes even infinite
loops like in the case of compound controls, like the spinbox, which
gives active focus to its textField content once it gains focus.
To fix, don't change the subFocusItem of the scene's root item unless
we have traversed up all the focus hierarchy.

Fixes: QTBUG-120542
Change-Id: Icf08edd2ff8b64662011f66c7137b44750cda9f2
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Shawn Rutledge <shawn.rutledge@qt.io>
(cherry picked from commit 1863c068191601beba0bcdc970185703c8296e3d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>
(cherry picked from commit 2f9ef932b4ddcaa32c51a60ce33dd4b6f27724c6)
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
(cherry picked from commit e529f8220dc0d50447a71fd788c50432e3793c9d)