ListView: allow binding on x and y in footer and header delegates
This is generally a bad idea, but it's technically a regression that was
introduced in f03a9839b689a3a1810846bb8ec08e02a9bf23b5.

The reason why it's not recommended to bind directly on x and y in
delegates, is because the delegate item positions are determined by the
item views layout, which are in conflict with the positional bindings.

However, it should be possible to partially support positional bindings
in the header and footer delegates, in ListView.
The reason for using the word "partially", is because a ListView with
horizontal orientation, should always place the header first (with x
typically being 0). But the y axis, can in theory be set directly, since
it won't cause overlap with the remaining delegate items.

Same idea for vertical ListViews. Except that the y axis become the
"primary" axis, when the orientation changes to vertical, meaning that
the x axis will then be safe to modify.

The model-view-delegate doc page is also updated, to discourage people
from binding directly on x and y, in delegates that are meant to be used
for item views.

Fixes: QTBUG-122665
Change-Id: I78363e159f14827a28dba7c72d1ecef45b0c6d2a
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>
(cherry picked from commit 27b40410613ad0ca753b7020527d03345c5aa8cf)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>