QQuickRhiItem: fix access specifiers of overrides
Overrides should not change the access specifier of the original
virtual method, because accessibility of the function should not
change depending on which concrete static type it's called on.

QQuickItem::isTextureProvider() and ::textureProvider() are public,
so the QQuickRhiItem overrides have to be public, too.

Found in API-review.

Change-Id: I4a9d4b8cd479eb8361a5b57e0fdd7ce3fdbbabcd
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>
Reviewed-by: Laszlo Agocs <laszlo.agocs@qt.io>
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>
(cherry picked from commit 95f77de0d45c6fba3465acf7b632e41cf10ed60d)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>