tst_QQuickTextDocument::overrideTextFormat() remove failing data set
The dataset "load html as autotext, switch to plain, back to auto"
caused permanent failures, because plain text escapes the <DOCTYPE>
header text and formatting tags, to mark them as metadata.
While the conversion back to HTML works, the conversion to autotext
keeps the escaping made in plain text.

A behavioral change would be necessary, to amend the conversion logic.
With the currently implemented logic, the specific test is bound to
fail.

Remove the test data set, to unblock submodule updates. Revisit the
issue at a later point.

Fixes: QTBUG-123014
Pick-to: 6.7
Change-Id: I0c5f461f17c4388dbd49796134e168a7afa0dd35
Reviewed-by: Oliver Eftevaag <oliver.eftevaag@qt.io>