SwipeDelegate: forward 'move' events to the child items
SwipeDelegate will currently forward both press and release events to
the child item under the mouse (swipe.left or swipe.right), but
not move events. If e.g swipe.left is a Button, the result is that
you can 'click' it, even if you release the mouse outside it's bounds.
The reason is that the button still thinks the mouse is on top of it,
because of the missing move events.

This patch will do the same as we do for both press and release (in
handleMousePressEvent() and handleMouseReleaseEvent()), and forward
the mouse events to the item under the mouse. This will make sure
that a Button (or MouseArea) will detect when it's no longer
hovered, which (in addition to updating e.g MouseArea.hovered
correctly), also ensure that we don't emit clicked when we shouldn't.

Pick-to: 6.7 6.6 6.5
Task-number: QTBUG-110475
Change-Id: I6d89e86e76c09168bebc12285c13c8bb68502bb9
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>