Fix attached property propagation through ListView with Behavior
If the user had code like this:

    ApplicationWindow {
        // ...
        Material.theme: Material.Dark
        // ...
        ListView {
            // ...
            header: Text {
                Behavior on Material.elevation {}

The meta type that QQuickAttachedPropertyPropagator saw when looking
for the attached parent of the Text item was e.g.
QQuickMaterialStyle_QML_125. However,
QQmlMetaType::attachedPropertiesFunc only has attached property data
for QQuickMaterialStyle (i.e. attached property types created from
C++), and so it would never find the attached Material object on the
parent window and would fall back to the engine's.

Fix it by finding the first C++ meta object, which works even for
attached types created in QML.

Fixes: QTBUG-122783
Pick-to: 6.5 6.7
Change-Id: Ifd1b887b514ade4f320ace8a033d54cd16dbee1e
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>