Remove the explicit declaration of QQmlComponent::createObject(QQmlV4Function *)
Now when manual moc provides the required QT_DISABLE_DEPRECATED_UP_TO
definition we don't need to declare the
QQmlComponent::createObject(QQmlV4Function *) for moc explicitly.

Also we need to manually add imported dependencies of Qt6::Qml to
make sure that we passed to moc all the required frameworks on the
Apple platforms.

Fixes: QTBUG-104898
Change-Id: I5404808b2589463d1b0ea5884f40e8e67cd5cb0e
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jari Helaakoski <jari.helaakoski@qt.io>
Reviewed-by: Ulf Hermann <ulf.hermann@qt.io>