MessageDialog: emit accepted, for buttons with roles Yes and Accept
After b00798df9d842570524550b6780a1debc4fe5387, the MessageDialog
would immediately close, when handling a click on a button inside
of it, and it would always emit rejected, since that was the default
result.
The issue was later fixed in 8dd68783642e6ea42e9a7b32e1671424d1d0eeed,
which was unfortunately not backported to 6.6 and 6.5.

This patch should serve as a very simple solution to fix the issue for
those that are using 6.6 and 6.5.

Fixes: QTBUG-122770
Pick-to: 6.5
Change-Id: Ia1f6a3fb0fd735188e74ad3e820fc3407f35c5cc
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>