ApplicationWindow: explicitly set background size if not explicitly set
41eb033fa2585b3644a545b75a414d4d3efa8b2c addressed QTBUG-118779 by
documenting that certain items like Image would need to have their
size explicitly set to the size of the window. This isn't in line with
the background-resizing behavior of Control, which explicitly sets the
size of the background if it hasn't been set (see
7e3c882d9a5cd6a25f547091a4ece786f516f2a4).

This patch partially reverts 41eb033fa2585b3644a545b75a414d4d3efa8b2c
by removing the change to the documentation. The relevant code in the
wearable example has since been changed, so there's nothing to revert
there.

[ChangeLog][Controls][Important Behavior Changes]
ApplicationWindow now explicitly sets the width and height of its
background if no size was explicitly set by the user. This matches the
behavior of Control, and ensures that if e.g. an Image is used as a
window's background, any changes in its implicit size (e.g. after a
source change) won't affect its actual size.

Fixes: QTBUG-121466
Pick-to: 6.5 6.6 6.7
Change-Id: I0fbe029626d736b4b96acce1a2ca450660a47271
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Richard Moe Gustavsen <richard.gustavsen@qt.io>