DomItem: allow fieldfilters in visitTree
Add a FieldFilter argument to DomItem::visitTree that allows to visit
only parts of a DomItem tree.

Add a test for it.

This is a preparation step for find usages in multiple files, as
visiting the unfiltered tree will take way too much time when
searching for usages.

Task-number: QTBUG-113827
Change-Id: I674dd65a52211a690f39cdae05b7b033482e5d5b
Reviewed-by: Fabian Kosmale <fabian.kosmale@qt.io>