QQuickMenu: respect Popup::position() when opening up a native menu
The popup() function can take a position, and it's also possible
to set the position using setPosition(). But the current
setNativeMenuVisible() implementation would ignore this, and
always use the mouse position instead.

This patch will remove this position override. The implementation
of popup(), that doesn't take a position as argument, will take
care of falling back to use the mouse position when no position
is specified.

Change-Id: I431630d9fd892843c0d00a5604650a9e23f7b6c1
Reviewed-by: Mitch Curtis <mitch.curtis@qt.io>
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>