Fix enum class lookup on Singletons
When we consolidated the code for looking up enums in QQmlType,
we exposed ourselves to a bug in initEnums: Not only CppType should be
affected by the logic for registering unscoped enums, but also C++
singletons (SingletonType).
Amends 4ceb343e7f006972940b6880f974770c02fa6b87 and
4e69f2468ad599b7175f60e08341d37689512c9c.

Fix this by consulting the meatobject, if there is one. Ideally, this
should probably stored in the QQmlType at type registration time, but
that can be cleaned up at a later point.

Fixes: QTBUG-123225
Task-number: QTBUG-123294
Pick-to: 6.7.0 dev
Change-Id: I7b09459ce4035acb2566b9550fa844897b1b363a
Reviewed-by: Sami Shalayel <sami.shalayel@qt.io>