Fix deadlock in long-running particle system
2^30 was used as a "big number" sentinel and assumed to
always be higher than the particle system's age. But
a particle system running for a bit over 12 days will
hit this limit. The result was that such particle
systems would never exit the loop in recycle() and
deadlock.

Instead we add the check for an empty heap to the
loop condition itself.

Pick-to: 5.15 6.2 6.5 6.6 6.7
Fixes: QTBUG-123111
Change-Id: I6bf9c5a27271e74f56728c0ad7bdabdd87e028b2
Reviewed-by: Volker Hilsheimer <volker.hilsheimer@qt.io>