CMake: Allow finding qmlimportscanner also in Qt build dir
The current implementation of _qt_internal_scan_qml_imports expects
to always find the qmlimportscanner path via its imported target.

In a follow up change, we want to enable using the
MACOS_BUNDLE_POST_BUILD option of qt6_generate_deploy_qml_app_script
for examples that are built in-tree as part of a -no-prefix build.

In this case, the imported target does not exist yet, and executing
the function will cause an error. To avoid that, change the code to
use the qmlimportscanner path from the build dir.

Make sure we still show the error in the cases where the tool must
exist at configure time:
- building a user project
- or trying to run qmlimportscanner at configure time while building
  examples in-tree.

Also re-adjust the error message to be more sensible.

Pick-to: 6.7
Task-number: QTBUG-90820
Task-number: QTBUG-96232
Task-number: QTBUG-102057
Change-Id: I0e73085ce1e53d2aee67493f51e148febf27c551
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>